Add Global Auto-Staging Option
Add a Toggle for automatically staging when the current stage's fuel is
depleted & the next stage has fuel. Currently the UI doesn't let you
toggle this off.

When running a mission, the UI will now fork an auto-staging thread that
connects with it's own RPC client, queries the current setting & fuel
status, and stages if necessary.